2015 Day 22
This one took SO LONG. I tried a bunch of different ways of doing it, including modeling the game as a tree of states, a graph of states, and finally by working through the game states in a priority queue. To be honest, I don't think any of these had an actual huge impact on the end result, but (a) now I have useful tree and graph libraries (and a priority queue and a heap, just in case), and (b) I learned a heck of a lot about data structures and profiling my code.

The culprit was the deep-equals library I had imported, which was EXPENSIVE. I imported it - against the general idea of me writing everything myself - because I was in the middle of writing one of the graph implementations and didn't want to put another yak on the stack. Well, 16 minutes turned into 4 seconds when I changed it out for a home-grown but less-exact version. Lesson learned.